requirements.txt
opencv-python
opencv-python-headless
ffmpeg-python
imageio-ffmpeg
seaborn
scikit-image
scikit-learn
scipy
h5py
PyYAML
numpy
matplotlib
lmdb
imageio
graphviz
fastjsonschema
pillow
tifffile
pyparsing
PyWavelets
networkx
yapf
tensorboard
tensorboardX
tqdm
munch
pytest
pytest-timeout
transformations
decorator
